kacom_model/requirements.txt
aiofiles==22.1.0
anyio==3.6.2
async-timeout==4.0.2
click==8.1.3
dnspython==2.2.1
ecdsa==0.18.0
fastapi==0.86.0
h11==0.14.0
httptools==0.5.0
idna==3.4
joblib==1.2.0
motor==3.1.1
numpy==1.23.4
pyasn1==0.5.0
pydantic==1.10.2
pymongo==4.3.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
redis==4.6.0
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
starlette==0.20.4
threadpoolctl==3.1.0
typing_extensions==4.4.0
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
